+++ lib/polyfill/v2_4/symbol.rb
@@ -1,0 +1,9 @@
+require 'polyfill/v2_4/symbol/instance'
+
+module Polyfill
+  module V2_4
+    module Symbol
+      include Instance
+    end
+  end
+end