+++ lib/polyfill/v2_4/symbol/instance.rb
@@ -1,0 +1,11 @@
+require 'polyfill/v2_4/symbol/instance/match_q'
+
+module Polyfill
+  module V2_4
+    module Symbol
+      module Instance
+        include MatchQ
+      end
+    end
+  end
+end