+++ CHANGELOG.md
@@ -3,11 +3,12 @@
 ## Added
 
  - v2.4 IO#lines
+ - v2.4 Regexp#match?
  - v2.4 String#each_line
  - v2.4 String#lines
  - v2.4 String#match?
  - v2.4 String#unpack1
- - v2.4 Regexp#match?
+ - v2.4 Symbol#match?
 
 # [0.3.0][] (2017-03-19)